backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
clip @ git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
ftfy==6.1.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
numpy==1.24.1
Pillow==9.3.0
pygame==2.1.2
pyrsistent==0.19.2
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
sqlparse==0.4.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
